Quick fix: couldn't remove question if the image == 'no image' or any image that does not exist in a folder where images are placed (set to '/static'. (FileNotFoundError). Also, couldn't remove question if the image == '' (FilePermissionError), because then it was trying to delete the '/static' directory.

+++ server.py
@@ -85,8 +85,12 @@ def delete_question(id):
     for question in all_questions:
         if question['id'] == id:
             image = question['image']
+            try:
+                if image != "":
+                    os.remove(f"static/{image}")
+            except FileNotFoundError as err:
+                print(err)
             all_questions.remove(question)
-            os.remove(f"static/{image}")
     index = 1
     for question in all_questions:
         question['id'] = str(index)
